Adding base class for perf, inter_token, and time_to_first latency records
