+++ model_analyzer/record/types/perf_latency_p90.py
@@ -16,11 +16,11 @@
 
 from functools import total_ordering
 
-from model_analyzer.record.record import DecreasingRecord
+from model_analyzer.record.types.perf_latency_base import PerfLatencyBase
 
 
 @total_ordering
-class PerfLatencyP90(DecreasingRecord):
+class PerfLatencyP90(PerfLatencyBase):
     """
     A record for perf_analyzer latency metric
     """
@@ -58,39 +58,3 @@ class PerfLatencyP90(DecreasingRecord):
         """
 
         return "p90 Latency (ms)"
-
-    def __eq__(self, other):
-        """
-        Allows checking for
-        equality between two records
-        """
-
-        return self.value() == other.value()
-
-    def __lt__(self, other):
-        """
-        Allows checking if
-        this record is less than
-        the other
-        """
-
-        return self.value() > other.value()
-
-    def __add__(self, other):
-        """
-        Allows adding two records together
-        to produce a brand new record.
-        """
-
-        return self.__class__(value=(self.value() + other.value()))
-
-    def __sub__(self, other):
-        """
-        Allows subbing two records together
-        to produce a brand new record.
-
-        ** Note this does reverse subtraction because
-            of the inverted nature of latency (lower is better)
-        """
-
-        return self.__class__(value=(other.value() - self.value()))
